Moved BoardView, SpaceView and BoardElementView to view.board package
